pyproject.toml
[project]
name = "delay-discounting-mvpa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "jmumford", email = "jmumford@stanford.edu" }
]
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.30.1",
    "ipython>=9.5.0",
    "nibabel>=5.3.2",
    "nilearn>=0.12.1",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.7.1",
    "scipy>=1.16.1",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.5",
]

[project.scripts]
delay-discounting-mvpa = "delay_discounting_mvpa:main"

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"


[tool.ruff]
extend-include = ["*.ipynb"]

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
extend-select = ["E501", "W", "N", "R", "I"]
ignore = ["F401"]
